refactor(spice): delete the old ChunkExecutorActor; coordinator is the only path
Phase 9 rollout: the per-shard ChunkExecutorCoordinator + PerShardExecutor
subsystem is now the sole spice chunk-execution path. The monolithic
ChunkExecutorActor (and the SPICE_PER_SHARD_EXECUTOR toggle) are gone.

- Relocate the shared pub(crate) executor helpers, ChunkExecutorConfig, and the
  ExecutorIncomingUnverifiedReceipts message into a new spice/executor_shared.rs.
- Delete chunk_executor_actor.rs: the actor struct/handlers/methods, the
  process_all_ready_blocks bootstrap, the ReceiptVerificationContext /
  VerifiedReceipts machinery, the NotReadyToApplyChunksReason /
  TryApplyChunksOutcome enums, ExecutorApplyChunksDone / FailedToApplyChunkError,
  and the testonly TestonlySyncChunkExecutorActor harness.
- Collapse the SPICE_PER_SHARD_EXECUTOR toggle in nearcore/src/lib.rs and
  test-loop setup.rs to always wire the coordinator (still gated by
  cfg!(feature = "protocol_feature_spice")); drop the const + chunk_executor_adapter.
- Wire the coordinator into jsonrpc-tests (mirrors nearcore/src/lib.rs).
- Disable the legacy TestEnv spice path: remove spice_chunk_executors /
  spice_execute_block and the produce_block hook (the integration-tests TestEnv
  framework is being migrated to test-loop). Spice execution coverage lives in
  the test-loop spice suite.
