+++ chain/client/src/spice/tests/mod.rs
@@ -1,3 +1,2 @@
-mod chunk_executor_actor;
 mod chunk_validator_actor;
 mod data_distributor_actor;
